add get_chat_list_item_by_id method
why? because desktop currently fetches the chatlist multiple times, even though it just needs the chatlistitem for one chat

Note: @r10s was worried before that exposing the method to get a single
updated chatlistitem could lead to race conditions
where the chatlist item is newer than the chatlist order.
But I don't think this will change anything for desktop besides making it a little faster.
(because currently desktop fetches the whole chatlist
instead of just the entry it needs when an entry updates)