Use NonZeroUsize
This avoids having to have a meaning for pause_guards_count == 0.  It
does however introduce a bunch of unwraps.  All of them but one are
safe:  When incrementing the pause_guards_count by one this could
panic if a lot of guards are taken out.  This is very unlikely in real
code though and probably means there's a bug elsewhere already.