Clear VerifiedOneOnOneChats setting on backup
If the user makes a backup from a UI that supports the experimental
verified 1:1 chats (e.g. nightly Android) and imports it into a UI that
doesn't, then this config should be cleared.

We already talked about this a long time ago, but I think we didn't
actually solve it back then?

Review with whitespace changes disabled.